chore: rename files in CreateGroup
