+++ src/App.tsx
@@ -7,16 +7,17 @@ import {
 } from "react-router-dom";
 
 import Layout from "./layout/Layout";
-import CreateCompletePage from "./pages/create/CreateCompletePage";
-import CreateDescriptionPage from "./pages/create/CreateDescriptionPage";
-import CreateLayout from "./pages/create/CreateLayout";
-import CreateNamePage from "./pages/create/CreateNamePage";
-import CreateNotionPage from "./pages/create/CreateNotionPage";
+
+import CreateGroupLayout from "./pages/create/CreateGroupLayout";
 import GroupDetailPage from "./pages/detail/DetailPage";
 import LoginPage from "./pages/login/LoginPage";
 import MainPage from "./pages/main/MainPage";
 import OnboardingPage from "./pages/onboarding/OnboardingPage";
 import Path from "./types/paths";
+import CreateGroupName from "./pages/create/pages/name/CreateGroupNamePage";
+import CreateGroupDescriptionPage from "./pages/create/pages/description/CreateGroupDescriptionPage";
+import CreateGroupNotionPage from "./pages/create/pages/notion/CreateGroupNotionPage";
+import CreateGroupCompletePage from "./pages/create/pages/complete/CreateGroupCompletePage";
 
 const router = createBrowserRouter(
   createRoutesFromElements(
@@ -26,18 +27,21 @@ const router = createBrowserRouter(
         <Route path={Path.Home} element={<MainPage />} />
         <Route path={Path.Group + ":uuid"} element={<GroupDetailPage />} />
 
-        <Route element={<CreateLayout />}>
+        <Route element={<CreateGroupLayout />}>
           <Route
             path={Path.Create}
             element={<Navigate to={Path.CreateName} />}
           />
-          <Route path={Path.CreateName} element={<CreateNamePage />} />
+          <Route path={Path.CreateName} element={<CreateGroupName />} />
           <Route
             path={Path.CreateDescription}
-            element={<CreateDescriptionPage />}
+            element={<CreateGroupDescriptionPage />}
           />
-          <Route path={Path.CreateNotion} element={<CreateNotionPage />} />
-          <Route path={Path.CreateComplete} element={<CreateCompletePage />} />
+          <Route path={Path.CreateNotion} element={<CreateGroupNotionPage />} />
+          <Route
+            path={Path.CreateComplete}
+            element={<CreateGroupCompletePage />}
+          />
         </Route>
 
         <Route path={Path.Onboarding} element={<OnboardingPage />} />
